+++ Core/Application/HostBuilderFactory.cs
@@ -3,7 +3,9 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Modding_Assistant.MVVM.Model;
+using Modding_Assistant.Core.Data.Interfaces;
+using Modding_Assistant.Core.Data.Models;
+using Modding_Assistant.Core.Data.Services;
 using Modding_Assistant.MVVM.Services.Implementations;
 using Modding_Assistant.MVVM.Services.Interfaces;
 using Modding_Assistant.MVVM.View.Dialogs;
