+++ Core/Data/Models/ModContext.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using Modding_Assistant.MVVM.Model;
 
-namespace Modding_Assistant.MVVM.Model
+namespace Modding_Assistant.Core.Data.Models
 {
     public class ModContext(DbContextOptions<ModContext> options) : DbContext(options)
     {
